Wrapper: Fix position attribute in LMS1xx scan data
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@19605 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: a6243c0c73d40d515e087dfb98893b0c90837299

+++ server/wrapper/lms1xx/de.uniol.inf.is.odysseus.wrapper.lms1xx/src/de/uniol/inf/is/odysseus/wrapper/lms1xx/model/Measurement.java
@@ -53,7 +53,7 @@ public class Measurement {
     private Sample[] samples16Bit;
     private Sample[] samples8Bit;
 
-    private float[][] position;
+    private double[][] position;
     private int positionRotationType;
 
     /**
@@ -210,14 +210,14 @@ public class Measurement {
      * @param position
      *            the position to set
      */
-    public void setPosition(float[][] position) {
+    public void setPosition(double[][] position) {
         this.position = position;
     }
 
     /**
      * @return the position
      */
-    public float[][] getPosition() {
+    public double[][] getPosition() {
         return this.position;
     }
